+++ spark/v3.2/spark/src/test/java/org/apache/iceberg/spark/source/TestStructuredStreamingRead3.java
@@ -25,7 +25,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.IntStream;
 import org.apache.iceberg.AssertHelpers;
 import org.apache.iceberg.BaseTable;
@@ -46,7 +45,6 @@ import org.apache.iceberg.relocated.com.google.common.collect.Iterables;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.apache.iceberg.spark.SparkCatalogTestBase;
 import org.apache.iceberg.spark.SparkReadOptions;
-import org.apache.spark.api.java.function.VoidFunction2;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Encoders;
 import org.apache.spark.sql.Row;
@@ -69,8 +67,6 @@ public final class TestStructuredStreamingRead3 extends SparkCatalogTestBase {
   }
 
   private Table table;
-
-  private final AtomicInteger microBatches = new AtomicInteger();
 
   /**
    * test data to be used by multiple writes each write creates a snapshot and writes a list of
@@ -117,7 +113,6 @@ public final class TestStructuredStreamingRead3 extends SparkCatalogTestBase {
             + "PARTITIONED BY (bucket(3, id))",
         tableName);
     this.table = validationCatalog.loadTable(tableIdent);
-    microBatches.set(0);
   }
 
   @After
@@ -153,51 +148,6 @@ public final class TestStructuredStreamingRead3 extends SparkCatalogTestBase {
 
     List<SimpleRecord> actual = rowsAvailable(query);
     Assertions.assertThat(actual).containsExactlyInAnyOrderElementsOf(Iterables.concat(expected));
-  }
-
-  @Test
-  public void testReadStreamOnIcebergTableWithMultipleSnapshots_WithNumberOfFiles_1()
-      throws Exception {
-    appendDataAsMultipleSnapshots(TEST_DATA_MULTIPLE_SNAPSHOTS);
-
-    Assert.assertEquals(
-        6, microBatchCount(ImmutableMap.of(SparkReadOptions.MAX_FILES_PER_MICRO_BATCH, "1")));
-  }
-
-  @Test
-  public void testReadStreamOnIcebergTableWithMultipleSnapshots_WithNumberOfFiles_2()
-      throws Exception {
-    appendDataAsMultipleSnapshots(TEST_DATA_MULTIPLE_SNAPSHOTS);
-
-    Assert.assertEquals(
-        3, microBatchCount(ImmutableMap.of(SparkReadOptions.MAX_FILES_PER_MICRO_BATCH, "2")));
-  }
-
-  @Test
-  public void testReadStreamOnIcebergTableWithMultipleSnapshots_WithNumberOfRows_1()
-      throws Exception {
-    appendDataAsMultipleSnapshots(TEST_DATA_MULTIPLE_SNAPSHOTS);
-
-    // only 1 micro-batch will be formed and we will read data partially
-    Assert.assertEquals(
-        1, microBatchCount(ImmutableMap.of(SparkReadOptions.MAX_ROWS_PER_MICRO_BATCH, "1")));
-
-    StreamingQuery query = startStream(SparkReadOptions.MAX_ROWS_PER_MICRO_BATCH, "1");
-
-    // check answer correctness only 1 record read the micro-batch will be stuck
-    List<SimpleRecord> actual = rowsAvailable(query);
-    Assertions.assertThat(actual)
-        .containsExactlyInAnyOrderElementsOf(
-            Lists.newArrayList(TEST_DATA_MULTIPLE_SNAPSHOTS.get(0).get(0)));
-  }
-
-  @Test
-  public void testReadStreamOnIcebergTableWithMultipleSnapshots_WithNumberOfRows_4()
-      throws Exception {
-    appendDataAsMultipleSnapshots(TEST_DATA_MULTIPLE_SNAPSHOTS);
-
-    Assert.assertEquals(
-        2, microBatchCount(ImmutableMap.of(SparkReadOptions.MAX_ROWS_PER_MICRO_BATCH, "4")));
   }
 
   @Test
@@ -528,8 +478,7 @@ public final class TestStructuredStreamingRead3 extends SparkCatalogTestBase {
    * appends each list as a Snapshot on the iceberg table at the given location. accepts a list of
    * lists - each list representing data per snapshot.
    */
-  private void appendDataAsMultipleSnapshots(List<List<SimpleRecord>> data)
-      throws InterruptedException {
+  private void appendDataAsMultipleSnapshots(List<List<SimpleRecord>> data) {
     for (List<SimpleRecord> l : data) {
       appendData(l);
     }
@@ -565,30 +514,12 @@ public final class TestStructuredStreamingRead3 extends SparkCatalogTestBase {
         .start();
   }
 
-  private int microBatchCount(Map<String, String> options) throws TimeoutException {
-    Dataset<Row> ds = spark.readStream().options(options).format("iceberg").load(tableName);
-
-    ds.writeStream()
-        .options(options)
-        .foreachBatch(
-            (VoidFunction2<Dataset<Row>, Long>)
-                (dataset, batchId) -> {
-                  microBatches.getAndIncrement();
-                })
-        .start()
-        .processAllAvailable();
-
-    stopStreams();
-    return microBatches.get();
-  }
-
   private StreamingQuery startStream() throws TimeoutException {
     return startStream(Collections.emptyMap());
   }
 
   private StreamingQuery startStream(String key, String value) throws TimeoutException {
-    return startStream(
-        ImmutableMap.of(key, value, SparkReadOptions.MAX_FILES_PER_MICRO_BATCH, "1"));
+    return startStream(ImmutableMap.of(key, value));
   }
 
   private List<SimpleRecord> rowsAvailable(StreamingQuery query) {
